Allowing trivial renames
Renaming (say) tape "T1" to "T1" used to be erroneous -- it would generate null, and also be flagged as an error in the interface.  But it should probably be a no-op instead; it's a valid expression of relational algebra so far as I know, it just doesn't do anything.  A few tweaks to the semantics make this a no-op instead of an error.  (In the future, we should probably also not bother constructing an expression for it in the first place.  But the semantic tweaks should remain; otherwise we can get some wrong answers w.r.t. tapes and such.)

This should come in handy when juggling tapes during replacing -- this way, "from X to X" and "from X to Y" can have the same grammatical structure even if some of the renamings involved are trivial.