Changed MatchFrom to support matching with more than two tapes.
Note: there appears to be a bug in MatchExpr such that output is not generated
correctly when there is an extra tape in the grammar not participating in the
match if the "first" Expr object in the child grammar is not for that extra tape.
For example MatchFrom t1, t2 of t1:h+t4:g generates no results, while
MatchFrom t1, t2 of t4:g+t1:h generates {t1:h, t2:h, t4:gg}, as would be expected.

This change is taken from the following commit on the (old) master branch,
but with the addition of more unit tests in testMatchFrom.ts:

0104ad45355f45ade30ee09010f65fb177c610c4
Darlene Stewart on 8/17/2021, 2:42:38 PM
Changed MatchFrom to support matching with more than two tapes.

Added unit tests for MatchFrom with two and three tapes (testMatchFrom.ts).

Fixed a bug in MatchState with the initialization of newBuffers that caused
matching the contents between more than two tapes to fail.