Framework for grammar-to-grammar transforms
We didn't previously have any particular way to transform grammars into other grammars.  That would, among other things, make it a lot easier to implement the renaming aspect of replacement.  (Rather than have ReplaceGrammar and/or Join handle the nitty gritty of renaming -- I started that and it's a slog, and hard to reason about -- we can selectively replace some of their children with the appropriate RenameGrammars, and let the specialist objects handle it.)

We *could* just have a XGrammar.replacementTransform() method that recursively transforms and returns the transformed grammar, but note that the vast majority of these would just apply the transformation to the children recursively and then return a clone of the original grammar with the transformed children instead.  There are only two-ish kinds of grammar objects that need a non-trivial implementation for this.

That's fine -- I mean, we have to write that code eventually somewhere -- but consider a likely future in which we have multiple different grammar-to-grammar transformations (e.g. optimizations, replacing the grammar as-written-by-the-programmer to a semantically equivalent grammar that will evaluate faster).  Most of those XGrammar.xTransform() methods will likewise be trivial; they'll be the same as our XGrammar.replacementTransform() methods except recursively calling .xTransform() on the children instead. So as our number of g2g transformations increase, each object has to get a new method for each one -- most of them trivial, and in aggregate hiding the few pieces of actually active code in pages of trivial and largely-identical code.

This situation is a good fit for the Gang-of-Four's "Visitor" pattern: when you have a recursive tree of component objects, and want to define additional algorithms on that tree without adding to the original objects' code.  Especially in our case, since we can envision that the behavior on most objects will be the trivial case; we can define a superclass that implements the identity transformation and then have further transformations subclass that, and only override the few methods that are non-trivial.

Visitor is a bit verbose in languages without multiple dispatch; that's why we need that XGrammar.acceptTransformation() method here.  Because we have to dispatch based on the class of two different objects, we need end up needing two method calls, one method inside each relevant class. (Multiple dispatch can be added to JS/TS and Python with libraries, but they're a bit overkill for our limited needs here.)

Also, note that our transformations are stateless -- they don't change anything about the original tree, they return a new tree.  (Some of the leaf nodes will happen to be shared, like we don't bother to clone EpsilonGrammars and such.)  It's become difficult to reason about state at this level of compilation, so being stateless should help here.  But also note that state calculated from the original tree won't necessarily be cloned over -- e.g. the .tapes property won't necessarily be cloned over.  (It shouldn't be if we're replacing grammars with RenameGrammars anyway -- the results will be different between the original and resulting trees!)