Replace rules from/to the same tape
It'll be common in replace rules that the programmer wants to "replace" material on the same tape.  (E.g., instead of saying "a on tape T1 corresponds to b on tape T2", saying "replace a->b on tape T1".)  From the point of view of the tape metaphor and its underlying implementation, this is "false" -- our abstract machine never backs up and replaces anything on a tape, for example.  These have to be two different tapes underneath.

So (like Hulden 2004) we handle this with renaming.

Originally I was trying to make ReplaceGrammar and JoinReplaceGrammar (a special version of JoinGrammar that understands how to join a list of replacements) handle the renaming internally, but this ended up being tricky.  Not impossible, but a pain to get right, and to keep in mind that what they were reporting (about tapes and such) wasn't necessarily true for all purposes.

Instead, I ended up using the transformation framework introduced in the last commit.  Rather than have these grammars act as if their children are renamed, we transform the tree such that the child grammars are wrapped in RenameGrammars.