Name resolution now stateless
Some of our grammar operations (like qualifyNames and calculateTapes) were stateful, they operated by adding state to the grammar tree.  Others were stateless (like the operation to adjust tape names in replacement rules) and would lose state -- sometimes appropriately (we don't want to keep the old, now-false tape calculation results), sometimes not (we ended up losing the name resolution results and had to redo them).

The more grammar operations that are stateless, the better we can compose/reorder/etc. them without having to reason about state.  This commit redoes name resolution so that it's stateless.  Instead of adding qualified names and such to NamespaceGrammar and EmbedGrammar objects, it:

(1) Changes the traversal algorithms to throw out the existing hierarchy of namespaces (which assign to unresolved names like "x") and replaces it with a new, single, flat, unnamed ("") namespace which assigns to resolved names like "nsName.x".

(2) Introduces an UnresolvedEmbedGrammar, which basically acts as an epsilon except for reporting its name as unqualified, and messaging an error if it survives until runChecks().

(3) The traversal algorithm attempts to resolve the names of UnresolvedEmbedGrammars, and if successful replaces them with EmbedGrammars.

(4) Previously, EmbedGrammars kept their resolved reference as state.  This can lead to problems if that reference has since been transformed into a new one.  We can't replace the references inside EmbedGrammars in one fell swoop while we make them, because they might transform after that EmbedGrammar itself is transformed.  So instead, we rebuild them with a reference to the new namespace.