Factoring out comp/gen into Interpreter
(1) The various compilation and generation code steps (especially those dealing with Grammars were repeated all over the place, and it was getting difficult to debug.  It used to be that every class could generate(), but as that started to get more an more steps, too many classes had that responsibility and too much code was repeated.

This is all factored out now and moved into the Interpreter class (which is merged with, and takes over from, the old top-level Gramble edifice).

(2) As mentioned in our 2021-10-14 meeting, we had been storing a partially compiled grammar -- specifically, the grammar created by TST.toGrammar().  But since then, a number of additional compilation steps have come up... and since we were only storing the grammar before those steps, we had to redo those steps every time we generated, sampled, ran checks, ran unit tests, made certain queries during testing, etc.  This was making samples and unit tests, in particular, take too long.

Now we store a grammar at roughly the last possible compilation step that would be shared between generation/sampling/checking/unit testing -- right after vocabulary is compiled/copied.

There are still some possible additional things to do after this, relevant to the particular action.  (Like a "parse" query that has restrictions, we have to figure out tapes, vocab, tape priority for the resulting FilterGrammar.)  But we don't expect the resulting grammars to need to be reused very often, so we treat that as ephemeral.

(3) There was a bug in the transformations w.r.t EmbedStates.  Basically, EmbedStates, as of the last commit, kept their namespace reference during a GrammarTransformation unless a specific .transformEmbed() override replaced that.  Without that specific override, it continues to refer to the old namespace.  However, since transformations clone, the namespaces after that are all new, and the embed is keeping a reference to the old one (and therefore looking up the old ones!)

So every transformation that clones (which is, at the moment, every transformation) would have to have a namespace-replacing .transformEmbed, and a namespace argument to represent the new namespace.  (This is simplified somewhat by our first transformation flattening all namespaces; without that this would be tricky.  But when transformation outputs are only going to have one namespace, so we only have to worry about passing in that one.)

So, (a) every transformation function needs a namespace argument, and (b) we build that basic .transformEmbed() switcheroo into the base class so that it becomes the default functionality of all new transformations.