Cleanup/refactoring of interpreter/transforms
(1) Interpreter.prepare() got the grammar ready for generating/sampling/testing/etc., and was called on demand.  But it was being called before nearly everything.  So now that code is just in the constructor.

(2) I had put in a compilationLevel enum into each grammar to keep track of how far we had gotten into Grammar transformation (for when we only go partway).  Now that we're doing those steps only once, in the Interpreter.constructor, those enums never do anything and can be removed.

(Exception, we DO do the name qualification one other place, in the loading loop.  But we end up throwing that result away anyway, it's just to see what symbol names are left unresolved.  Maybe we shouldn't throw those away, but I don't think it wastes very much time; I should check sometime.)

(3) Now that we've moved compilation/generation/unit testing/etc. out of Grammars, they no longer call any GrammarTransforms on themselves.  The only thing they do with GrammarTransform is to .accept() them.  So they no longer have to know about any concrete GrammarTransforms; they only need to know about that interface.  So the concrete implementations are now moved into their own file, transforms.ts.

(4) As mentioned in the last commit, we now store the grammar object much closer to the end of compilation.  But there are still a few operations remaining, like collecting new vocab when we're actually generating from a Filter.  (We do that when it's a restricted generation, and in unit tests.)  Previously, we were .collectVocab() and .copyVocab()-ing through the entire grammar again, but we really only have to .collectVocab() from the new material.