CharSet grammars and exprs
Implementing a regex-style charset (e.g. [abcde]).  This has a different effect from doing a union like a|b|c|d|e; this puts a single token with multiple bits set into the output, rather than multiple token possibilities each with a single bit set.

We don't have an immediate programmer-facing use of this, but it's handy for testing, because it gives us the ability to construct particular bit sets without using negation (because Negation Is Weird) and without using dots and vocab manipulation.

In particular, I used this to express tests that will succeed under Darlene's current implementation of Matching, but would fail if we implemented Pat's incorrect suggestion on 2021-10-13 (or anything that shares the problem where an earlier set is narrowed upon a subsequent match on another tape).  Should this problem arise in Matching, it'll be easier to narrow down that it's really Matching causing the problem, rather than it being a negation or vocab problem.