Uniqueness params for unit tests
Some of our ported projects (like Kawennonnis) needed a slightly more sophisticated unit testing framework.  Rather than just confirm that {T1:a, T2:b, t3:c} is a generated output, the unit tests suppose that you can say,
"Given {T1:a, T2:b}, the *only* thing T3 can have is "c"."

For some time, this kind of test wasn't possible, because there was no way to distinguish that columns of a unit test had different roles in that test.  However, the parameter tag framework (introduced for replacement rules) gives us that.

So now you can say:

   test:, T1, T2, unique T3
        , a , b,  c

And if there's any result including {T1:a, T2:b} but including anything but "c" on T3, it's an error.  So {T1:a, T2:b, T3:z} is an error, and so is {T1:a, T2:b}.

Also, this commit expands some of the syntax error testing for replacement params (for example, making sure that this unique keyword is invalid in those param dicts).  There's also some code cleanup here and there.