Unicode tokenization sensitive to diacritics
Formerly, strings containing Unicode diacritics and similar symbols could cause surprises.  Vowels with diacritics, for example, could satisfy a startswith condition with the plain vowel, depending on whether they were represented in Unicode with a combined character or a letter with a mark -- the former would not satisfy the startswith, the latter would!

This commit introduces predictable results for handling of Unicode diacritics and similar:

(1) All strings are converted to NFD first, so there's no difference whether the programmer chose a combined character or a letter and mark.
(2) Diacritical marks and modifier letters (more specifically, all characters in the classes 'Lm', 'Sk', 'Mc', 'Me', and 'Mn') are treated as belonging to the previous token.
(3) The IPA "tie" characters (the combining double breves) also include [what would be] the following token in the previous token.  For example, 't͡s' would be tokenized as a single token -- the 's' gets added on to the token because it's preceded by a tie.  This also takes into account modifiers; the 's' will get added on even if preceded or followed by a diacritic/modifier.

This commit also commits some files from the previous commit that were accidentally not git added.