Better CSV parsing; CLI reintegration
(1) Better CSV parsing, to handle things like commas and quotes inside cells.

(2) Hooked everything back up to the command-line interface, and added a command-line argument to specify JSON or CSV format.  CSV is now the default.

(3) Added a "verbose" parameter to the Interpreter; right now that just outputs elapsed time logs. (The only part of compilation that's noticeably slow, by the way, is vocabulary collection. I can't think of an obvious way to speed that up.)

(4) We no longer bother to add characters to hidden tape outputs.  It doesn't currently make a difference (we're not using a lot of hidden tapes) but it may when we do a lot of replacement.

(4b) Also, adding characters to hidden tapes somewhat screwed with character counting -- hidden tapes are not supposed to be part of the "public interface" of a module, and should be invisible to the user, but if the client says "Max 100 characters" and all of those characters are spent on hidden tapes, that's a situation where it really does affect the interpretation of the module.  So now only characters on non-hidden tapes count towards the character maximum.  Maybe this is a bad idea, because the point of max characters is to avoid infinite loops and such and maybe there could be that on hidden tapes, but if it comes to that we can reverse this.