Re-enabled sampling in CLI
(1) The CLI does sampling now, in the same manner as generation and using mostly the same arguments.

(2) To support this, there's a new method sampleStream() in the Interface class, which (like generateStream()) returns a generator rather than a list.

(3) The index files now export timeIt() for convenience, so that clients don't have to reimplement time logs.

(4) Small cleanups here and there, like changing "var" to "let" in the GSuite backend code.