Small optimizations and cleanup
Some small oversights and inefficiencies were slowing us down slightly; removing them speeds us up by about 10%.

(1) For some reason I was still shuffling the next states in depth-first search (as we do for sampling); I think I added that to test something and forgot to take it out.

(2) It's faster to x.push(...y) than x = x.concat(y)

(3) Generating or sampling with a restriction CAN change the grammar, so we have to re-targeting, re-taping, and re-vocabbing.  However, we were doing this inside the sampling loop (inside which, the restriction is never going to change).  So now we do that outside the sampling loop.

Also just cleaned up some dead branches, and changed the while expression in the generation methods to avoid an annoying check we had to do only for linting.