Improved symbol assignment in TSTs
Implementing a new algorithm for assigning grammars to namespaces in TSTs; now we recurse through TST enclosures using assignToNamespace().  Actually, this is resurrecting an old algorithm.  But unlike the old one, it takes into account last-children of namespaces and can thus better detect when overt assignment is/isn't necessary.

This makes it easier to catch certain kinds of errors in which the programmer makes a syntactically possible configuration of TstEnclosures, but in which a grammar ends up lost because it doesn't assign to anything.

It also makes it possible to put assignments elsewhere, other than as TstSheet children.  (We'll need this for programmer-defined namespaces in the near future.)  Formerly, we only created TstAssignments when an unknown op occurred in the first column; otherwise, an unknown op was an error.  Now, every unknown op is a TstAssignment, but because we can now detect syntactically inappropriate assignments, we can catch the error that way.  (For example, an assignment op as an argument to a binary op is meaningless; we can now detect exactly that, rather than avoiding the situation by saying "only first column ops can be assignments".)