More assignment improvements
The previous commit wasn't quite right in all details (e.g. it was allowing assignment in the "sibling" argument of a binary op).  This refines the assignment rules to be more specific to the particular kind of TstEnclosure.

Unlike other binary ops, unit tests DO allow you to assign in their first ("sibling") argument.  This is because unit tests are semantically transparent; they don't change anything, so there's no real harm in letter them sit directly beneath the assignment they're testing, rather than have to be its child.  So for example, you're now allowed to do this:

VERB:, embed, embed
     , stem,  suffix

test:, text, gloss
     , foobar, run-1SG

You're not required to do this, you can just do it the old way, but there was no syntactic/semantic reason to disallow this.

This commit also restores the negation tests that were accidentally deleted in the last commit.