Fix when "replace:" op is missing child
In the process of adding unit tests for errors that we check but never happened to construct unit tests for, I found a doozy.

When an operator is missing one of its operands, that operand is treated as an epsilon.  However, epsilons have no tapes.  When you join an epsilon with anything, that anything just ends up being itself (because they have no tapes in common).

So when the "replace:" op is missing its sibling argument (the grammar it's doing a replacemnet on), we end up doing unconditional generation from the replacement, which is infinite.

This is semantically correct, but not the results anyone actually wants to see when they make a mistake such that the sibling argument is epsilon, or another mistake where the sibling doesn't happen to have that tape.

Since a wide variety of syntax errors could potentially cause this, TSTs aren't the right time to catch this.  The best time to catch this is during the tape replacement stage -- that's when we work out what tape exactly has to be renamed in that sibling argument.  Now, if the sibling doesn't even have the appropriate tape, we ignore the replacement altogether -- instead of joining it with the renamed sibling, we just return the sibling.