More thorough testing
Added some additional tests for syntax errors in TstUnitTests (like last time, errors that we checked but weren't testing).  Changed the checking a bit so that it's now okay to nest a table: op under a test: op.  (There's no reason to for a programmer to do this, but there's no syntactic/semantic reason for us to forbid it, under the principle "table: should be a no-op".)

Also added generation tests for various old syntax error tests.  (That is, some of our oldest tests for sheet syntax errors didn't generate.)  The problem found in the previous commit illustrates the importance of actually generating them: sometimes our attempts at repairing the error lead to infinite generation or other kinds of outputs that won't be helpful.

In doing this, I discovered that an ill-formed header doesn't just epsilon-ize its OWN content cells, but accidentally epsilonized all the cells to the left of those, too.  That's not a helpful repair! (Headers construct their grammars not just from their content cells, but also the grammar that results from the cell(s) to the left.  Invalid headers were just returning epsilon, regardless of what left was!  Now they just return the results of left.)