Small improvements/fixes to TSTs
(1) TstSheet and TstProject are greatly simplified.  (a) TstSheet is now TstNamespace, and it makes better use of its inherited TstEnclosure methods to become very simple.  Really all it's doing is acting differently w.r.t. assignToNamespace().  (b) TstProjects are now just a subclass of TstNamespace that handles adding children differently.

(2) While nothing actually goes *wrong* in the syntax/semantics to name your sheets reserved words (like "maybe.csv"), this should be an error regardless.  Names of sheets are also symbols -- programmers can refer to them -- and it's better to have a general rule "Symbols can't be reserved words", even if in *this* case nothing actually goes wrong.  (It doesn't go wrong because there's no ambiguity in parsing; it's not like sheets are named in a cell that the parser is iterating over, such that they could be confused with an operator.)  I changed the tests that accidentally did this, and added a test that does this on purpose.

(3) In a recent commit, we simplified some of the Tst logic by making non-existant siblings/children TstEmpties rather than undefined (similar to making grammars be EpsilonGrammars rather than undefined).  This extends that pattern to the first default child of TstRows, which simplifies some of the TstRow and TstHeadedCell logic because they can now count on that existing.