Many bugfixes to cell & header parsing
Found many bugs in cell and header parsing, some of them major.

(1) A while back, I announced that empty cells had a new semantics (they were now empty literals rather than not-parsed-at-all), and that this would affect filters (filtering with an empty literal is quite different than not filtering).  We agreed that this was probably good, because previously there hadn't been a way to filter on the empty string, and that's something reasonable to do sometimes.

But it turns out that you couldn't actually do that in the tabular syntax.  The code that creates filters (in the superclass LogicHeader) was still checking to see if the cell text was empty, and if so it was not creating a filter at all.

Fixing this revealed a case of cell parsing (and mini-parsing in general) that we hadn't covered: none of our possible mini-parser grammars could accept an empty string as valid, there was just no parser-combinator that didn't try to consume at least one token.  We initially made this for header parsing, after all, and the empty string is no kind of header at all, nor a valid part of any header.  But the empty string is a valid cell value now!

So now there's a parser-combinator for the empty language, MPEmpty<T>(), an empty result for cell parsing (CPEmpty), and the LogicHeader logic has been changed to handle it.

Btw, if you need access to an Empty in the cell syntax, you can get it with "()".  (You can't just go "1SG|", that's an error. I think it's more likely that a programmer deletes a condition by accident and forgets to remove the | than really means the empty.  So if you want that, you have to go "1SG|()".)

(2) Negation in cell parsing was too high in precedence, leading to things like "~1SG|2SG" == "~(1SG|2SG)".  That's not how I mentally parse that.  So now "~1SG|2SG" == "(~1SG)|2SG".

(3) We weren't testing for thrown errors properly; this fixes that.

(4) We were trying to test for thrown errors by parseHeader(), but it no longer throws errors; instead it constructs an ErrorHeader or subclass of that.  So now we test that the result is an instance of that, rather than test for the thrown error.

(5) We were never actually catching the errors thrown by parseBooleanCell(); a malformed filter like "v~" would crash the whole thing.  Now we catch them and send an error message to the interface.