Tests for complex filter headers
Added a few tests to test more complex filter (equals etc.) headers, like how they interact with slash headers.

The way things are currently coded, alternation as expressed by | in a cell scopes over concatenation as expressed by a / in the header.  (So if you say "equals X/Y:a|b", {X:a,Y:a} and {X:b,Y:b} are good but {X:a,Y:b} and {X:b,Y:a} are filtered out.)  Intuitively I would kind of expect the opposite scope.  But this is a weird thing to express in any case; the purpose of the test is more to check whether it has ANY result, doesn't crash, etc.