Regex headers
(1) In an early gramble, we had a header "alt X" that turned on the "|" operator, and "@X" partially did the same (and also caused a join).  But @X had been turned off.  I had quietly turned this back on with a header "logic X", but hadn't used or tested it.

This commit officially "turns it back on", introduces tests, and renames it "re" for regular expression.  (It's not yet a regular expression, it's still the same limited logic as before, but it will eventually handle simple regexes as well.)

(2) Escaping special characters under re, equals, startswith, etc. now works.  Previously there had been no way to express (for example) a literal | under these headers, now you can just do "\|".

(3) Fixed a few more non-working expect().to.throw() tests.

(4) We hadn't been testing multiple applications of replacement in the tabular syntax yet; added those.