Support for sequences in re/filter cells
(1) Previously, in cells underneath re/equals/startswith/etc. cells, you could say things like "(1SG|2SG)", but you couldn't say "(1|2)SG" -- there was no mechanism to parse it or turn that parse into a SequenceGrammar.  It could only be one thing -- an alternation, a negation, a literal, or (just recently) empty.  Now it can be a sequence too.

That changes the grammar of cellparsing significantly, and required a new MPParser, MPRepetition, which lets us define a possibly infinite sequence of X.  (Note that when these become CPResult objects they're CPSequences, and when those become grammars they're SequenceGrammars, but the construction in the language that *describes* them is a repetition.  There's no necessary connection between the types, just like CPAlternation and AlternationGrammar are specified by a MPSequence.)

Since MPRepetition can possibly include zero repetitions, that eliminates the need for MPEmpty and thus CPEmpty.  MPEmpty is just the zero case of MPRepetition, and CPEmpty is just the zero case of CPSequence.

(2) Reworked the error behavior of cell.ts and headers.ts so that they work in the same way.  The miniParser no longer throws exceptions, it returns the list of results no matter how many there are or should be.  cell.ts and header.ts now do the counts and make sure they're equal to 1.  If there are zero, this is probably a programmer error and results in a special placeholder, CPError and ErrorHeader respectively.  But if there aren't zero, this ISN'T a programmer error, this must be a Pat error in not specifying the respective grammar in a way that it's unambiguous.  So that DOES throw an exception in both cases.

(3) CPSequence makes the structure of CPResults a lot more complicated, and it became really gnarly to test that particular parse results had the structure and types they're supposed to have.  So instead of trying to grab & test particular children (e.g., that X is a CPSequence, that it has 2 children, that the first child is a CPAlternation, etc.), we just ask for an ID of that object that's calculated recursively, like "[OR[1SG,2SG]]".

This deliberately doesn't look like the regex syntax so that when something goes wrong (like failing to parse a | as an alternation) it's more obvious.

(4) Made tests for sequences under re and equals headers.