Repetition operators for re/filter cells
This commit adds *, ?, and + to the operations that are valid in cells underneath re, equals, startswith, etc. headers.

With this, we now support Kleene regular languages with negation in such cells.  Note that this is a different set of languages than is supported by "regexes" in Python/Perl/etc.; those are extended regular expressions with a different set of operations and languages that can be described.  (For example, there is no backreference here, whereas in those languages' regexes, negation is highly restricted.)