Renaming cell-parsing to Regex for clarity
We use the word "Cell" to mean a lot of things in this project.  One thing that never quite had a good name was the constellation of objects/methods devoted to parsing special operators in "content" cells (not ops, not headers, just cells in the context we would usually find literals, but aren't literals).  That was just "cell parsing", resulting in CPResult (cell parsing result) objects, which isn't very descriptive.  They weren't regexes because they didn't support enough operators to describe regular languages.

But now they do, so it's correct and clear to call those objects Regexes.  This renames those objects, and files that deal with them (like testCellParser.ts to testRegex.ts).