Bugfix for replacement+namespaces
(1) An insidious bug resulted from combining replacement and a non-trivial namespace, that potentially can cause infinite generation.

What ended up being the culprit was two things:

(1) JoinReplaceGrammar not prioritizing its intersection the way JoinGrammar does.  I wasn't convinced it had to, but it does.  (JoinReplaceGrammar now transforms to a JoinGrammar in ReplaceAdjuster, though, so that handles that.)

(2) Interpreter was asking for tapePriority from the root of the grammar (in this case a namespace), but namespaces' tape priorities have no particular meaning; there's no reason to believe it should be a good priority for any given symbol in it.  We should have been asking for the tapePriority from the relevant symbol.

This gave a bad tape priority for the join-replace combination, asking it to generate on the output tape rather than the tape the two sides share.  But you can generate from the output tape indefinitely and never end up checking the shared tape to see whether anything can be matched on it.

(2) This commit also makes some changes to simplify Expr.id, so that it's easier to read when debugging -- particularly by taking sequences of T:a+T:b+T:c and attempting to simplify them to T:abc.

(3) To simplify the construction of the abbreviations above, I also switched ReplaceGrammar to use createDotRep() and made a special case for that in the infinite case to generate a DotStarExpr.  DotStarExpr is semantically equivalent to an infinite repetition of a dot, but since we do that so often, making them one object lets us more conveniently handle things like the abbreviation above.