Filter headers are no longer automatically regexes
(1) As agreed upon in our meeting, now that there's a "re" header, programmers should be explicit in filters ("equals" etc.) about whether to interpret the content below as a regex or not.

(2) "startswith" and "endswith" are now "starts" and "ends" for brevity.  (Now that there's another header in the string, e.g. "starts re text", brevity is more important.)

(3) Added tests for the combination of "re" and "embed".

(4) "Re" and slash headers were leading to ambiguity; turns out "re" was too far down in the header grammar.  But this was a good opportunity to:

(5) "re" with a slash header now results in an error.  I was already a  hesitant about the way re and slash scoped, and it's a weird thing to do in the first place, so let's just say it's not allowed.  (Also, we're eventually probably going to have a rule that anything under "re", even an embed, can only have one tape, so this would be consistent with that.)