Small QoL/API/testing improvements for headers
(1) In the sheets interface, regex and literal content are now distinguished by font color -- content that will be interpreted as a regex is a slight red-purple tint as a subtle reinforcement and warning.  If you're thinking this is a regex but it's not red, something's wrong; if you're thinking it's not a regex but it's red, something's wrong.  Literal content is a slight green-blue.

(2) Headers now have an .id property, which lets us unit test them the same way we did regexes -- not by going through the structure and by checking the types of children (which is awkward in mocha+TS) but just by saying "does the parse result have this particular .id string?)  This makes unit testing complex headers much easier.

(3) Interpreter.getTapeNames() used to return a list of [tapeName, color] pairs.  But currently, most calls of this don't need the color information, leading to annoying .map operations to grab the tape name alone.  Now .getTapeNames() just returns the tape names, and there's a method .getTapeColor(tapeName) that will calculate the appropriate color as needed.