Right-to-left generation
Only three Exprs have an inherent directionality: LiteralExpr, ConcatExpr, and StarExpr (but for the latter, only because its deriv is a ConcatExpr).  This commit implements versions of these that work in the opposite direction -- from the end to the beginning.

It ends up being somewhat faster on our big test grammar (although that might just be because that uses "ends", a grammar that uses "starts" might be slower).  So in the future, we should make this a parameter somewhere, and attempt to guess it based on the grammar.  (Right now, it's kind of specified all over the place, to switch between LTR and RTL requires a bunch of little changes.)

I had hoped it'd be possible to make this choice on a tape-by-tape basis, but it won't be, not with our current implementations.  A lot of the differences between directions are in deriv, making it seem like direction could just be a parameter to the current call to deriv, associated with the tape... but it's not just there, it actually affects how you build the Expr tree.  Some configurations are incorrect for the other order, others are just slow.  You would end up building the derivative tree for one direction, and then on another tape you'd get an incorrect answer or an inefficent one, because the tree was built in the wrong way.  (Making a tree that works both ways wouldn't be impossible, but it's not a simple change, you'd have to implement completely new ways of doing concat/star/etc. with this in mind.)

This commit also switches our algorithm to be based on concreteDeriv() (this was implemented in the previous commit but we didn't yet switch to it).  It also eliminates dictDeriv() -- having two kinds of deriv around was okay, but three was getting unwieldy.