New MatchState implementation
We had two difficulties with MatchState:

(1) If you set up a Match between T1 and T2 where T1 was more restrictive than T2 (e.g. Match(t1(h)+t2(.*))), but then the grammar as a whole caused T2 to be generated first, you could get effectively-infinite generation on T2, even though everything but "h" would, if we consulted t1(h), immediately fail.

(2) Material on unrelated tapes (tapes relevant to the child of Match but not a tape we've asserted a match on) isn't always consumed, leading to failed tests, and it depends on the order that we evaluate in.

Although both of these are order dependent, they were probably unrelated -- (1) was a problem in general, but (2) was (I think) specifically a problem in MatchState.delta().  (I think the way MatchState was implemented, it made certain corner cases of .delta() hard to implement, and I was lazy.)  However, the new implementation of MatchState that addresses (1) also makes it easy to handle the corner cases of .delta().

I push this to a new branch, however, because it doesn't seem to play nice with joining replacement rules to grammars -- all rules become optional and it's still unclear why.  We pass all the existing match/replace/etc. tests, but there must be a behavioral difference from the old MatchState that we're not capturing in tests, because the new one acts differently under join.