Added otherContext to Replace and ReplaceGrammar, as well as maxCopyChars.
otherContext identifies context to match on other tapes other than the fromTape
(identified by fromState) and toTape (identified by toState).

maxCopyChars is a character limiter for copy-through when replacement is not
applicable. This is needed for unit tests because copy-through is done with a
global match instead of a repetition of per-instance matches. maxCopyChars
defaults to the value of maxExtraChars so that existing unit tests all work as
before. A larger value should be specified in tests where one wants to test
longer copy-through text, but be careful not to make it generate effectively
forever because copy-through is essentially a .* match.