Found another churn-forever configuration
This one is probably the simplest so far: joining T1:whatever with a match of .* on both T1 and T2, then manipulating tape priority so that T2 is evaluated first.  This is unrelated to past and future bugs in matches/replaces/etc.; rather, it suggests that the whole idea of "finish tape X completely before ever moving on to Y" is untenable, and that tape rotation is a necessity during generation.

I also moved all such churn-effectively-forever tests into a special test file, testMatchDanger.ts, so that it's easier to turn them all off and on.  (Because churn-effectively-forever tests don't just fail, they churn forever.  If you're working on fixing something else, you have to turn them off.)